translate-cpp: split ground-fact types out of condition.h
condition.h had grown to 602 lines mixing three unrelated concerns: the lifted
Condition AST, and the ground-fact *runtime* types used by instantiation and
STRIPS->SAS translation (VarMapping, GroundKey/GroundKeyHash, FactId, FactMap,
GroundLiteral, GroundEffect). Move the latter into a new pddl/ground_facts.h;
condition.h includes it (its instantiate() signatures reference those types), so
no other include changes are needed. condition.h is now just the AST hierarchy.

Pure type relocation, no logic change. Byte-identical: 242/242 equivalence runs
(every family x every option).
